// src/3d/scene/bibim-result/model/bibim-result-scene.types.ts
export type BibimResultSceneSlice = {
  color: string;
  diffuseMapUrl?: string;
  displacementMapUrl?: string;
  id: string;
  normalMapUrl?: string;
  value: number;
};

export type BibimResultSceneSliceMesh = {
  color: string;
  diffuseMapUrl?: string;
  displacementMapUrl?: string;
  id: string;
  normalMapUrl?: string;
  thetaLength: number;
  thetaStart: number;
};